+++ interapp-frontend/src/utils/index.ts
@@ -1,4 +1,4 @@
-export * from './parseErrorMessage';
+export * from './parseServerError';
 export * from './getAllUsernames';
 export * from './remapAssetUrl';
 export * from './wildcardMatcher';
